added collector for intent label
